Fixing types.

+++ stlearn/spatials/clustering/localization.py
@@ -8,7 +8,7 @@ from sklearn.cluster import DBSCAN
 def localization(
     adata: AnnData,
     use_label: str = "louvain",
-    eps: int = 20,
+    eps: float = 20,
     min_samples: int = 0,
     copy: bool = False,
 ) -> AnnData | None:
